backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
motor==3.3.2
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
websockets==12.0
aiofiles==23.2.1
structlog==23.2.0
slowapi==0.1.9
python-dotenv==1.0.0